+++ this/tests/fast_mapper.rs
@@ -4,12 +4,29 @@ mod tests {
     use this::fast_mapper::FastMapper;
     //use std::path::Path;
     use kmers::naive_impl::Kmer;
+    use this::mapping_info::MappingInfo;
+    use this::ofiles::Ofiles;
+    use std::fs::File;
+    use std::path::PathBuf;
+
     #[test]
     fn check_geneids() {
         let mut mapper = FastMapper::new( 16 );
 
         let mut geneid = 0;
-        
+        let outpath = String::from("testData/");
+        let log_file_str = PathBuf::from(&outpath).join(
+            "Mapping_log.txt"
+        );
+        let log_file = match File::create( log_file_str ){
+            Ok(file) => file,
+            Err(err) => {
+                panic!("Error: {err:#?}" );
+            }
+        };
+        let ofile = Ofiles::new( 1, "Umapped_with_cellID", "R2.fastq.gz", "R1.fastq.gz",  outpath.as_str() );
+        let mut results = MappingInfo::new( log_file, 0.20, 1232423523523, ofile );
+
         mapper.add( &b"ATCCCATCCTTCATTGTTCGCCTGGA".to_vec(), "Gene1".to_string() );
         mapper.names4sparse.insert( "Gene1".to_string(), geneid );
 
@@ -22,14 +39,14 @@ mod tests {
         geneid +=1;
         mapper.names4sparse.insert( "Gene1".to_string(), geneid );
 
-        assert_eq!( mapper.with_data, 8 );
+        assert_eq!( mapper.with_data, 31 );
 
-        assert_eq!(  mapper.get( b"ATCCCATCCTTCATTGTTCGCCTGGACTCTCAGAAGCACATCGACTTCTCCCTCCGTTCTCCTTATGGCGGCGGC" ), Some(0) );
-        assert_eq!(  mapper.get( b"CGATTACTTCTGTTCCATCGCCCACACCTTTGAACCCTAGGGCTGGGTTGAACATCTTCTGTCTCCTAGGTCTGC" ), Some(1));
+        assert_eq!(  mapper.get( b"ATCCCATCCTTCATTGTTCGCCTGGACTCTCAGAAGCACATCGACTTCTCCCTCCGTTCTCCTTATGGCGGCGGC", &mut results ), Some(0) );
+        assert_eq!(  mapper.get( b"CGATTACTTCTGTTCCATCGCCCACACCTTTGAACCCTAGGGCTGGGTTGAACATCTTCTGTCTCCTAGGTCTGC", &mut results ), Some(1));
 
         mapper.add( &b"CGATTACTTCTGTTCCATCGCCCACACCTTTGAACCCTAGGGCTGGGTTGAACATCTTCTGTCTCCTAGGTCTGC".to_vec(), "Gene3".to_string() );
 
-        assert_eq!(  mapper.get( b"CGATTACTTCTGTTCCATCGCCCACACCTTTGAACCCTAGGGCTGGGTTGAACATCTTCTGTCTCCTAGGTCTGC" ), Some(1) );
+        assert_eq!(  mapper.get( b"CGATTACTTCTGTTCCATCGCCCACACCTTTGAACCCTAGGGCTGGGTTGAACATCTTCTGTCTCCTAGGTCTGC", &mut results ), None );
 
         let mut gnames = Vec::<String>::with_capacity(3);
         gnames.push( "Gene1".to_string() );
@@ -44,6 +61,20 @@ mod tests {
     fn check_write_index() {
         let mut mapper = FastMapper::new( 16 );
 
+        //log_writer:File, min_quality:f32, max_reads:usize, ofile:Ofiles,
+        let outpath = String::from("testData/");
+        let log_file_str = PathBuf::from(&outpath).join(
+            "Mapping_log.txt"
+        );
+        let log_file = match File::create( log_file_str ){
+            Ok(file) => file,
+            Err(err) => {
+                panic!("Error: {err:#?}" );
+            }
+        };
+        let ofile = Ofiles::new( 1, "Umapped_with_cellID", "R2.fastq.gz", "R1.fastq.gz",  outpath.as_str() );
+        let mut results = MappingInfo::new( log_file, 0.20 as f32, 1232423523523, ofile );
+
         let mut geneid = 0;
         
         mapper.add( &b"ATCCCATCCTTCATTGTTCGCCTGGA".to_vec(), "Gene1".to_string() );
@@ -53,10 +84,10 @@ mod tests {
         geneid +=1;
         mapper.names4sparse.insert( "Gene1".to_string(), geneid );
 
-        assert_eq!( mapper.with_data, 8);
+        assert_eq!( mapper.with_data, 31);
 
-        assert_eq!(  mapper.get( b"ATCCCATCCTTCATTGTTCGCCTGGACTCTCAGAAGCACATCGACTTCTCCCTCCGTTCTCCTTATGGCGGCGGC" ), Some(0) );
-        assert_eq!(  mapper.get( b"CGATTACTTCTGTTCCATCGCCCACACCTTTGAACCCTAGGGCTGGGTTGAACATCTTCTGTCTCCTAGGTCTGC" ), Some(1) );
+        assert_eq!(  mapper.get( b"ATCCCATCCTTCATTGTTCGCCTGGACTCTCAGAAGCACATCGACTTCTCCCTCCGTTCTCCTTATGGCGGCGGC", &mut results ), Some(0) );
+        assert_eq!(  mapper.get( b"CGATTACTTCTGTTCCATCGCCCACACCTTTGAACCCTAGGGCTGGGTTGAACATCTTCTGTCTCCTAGGTCTGC", &mut results ), Some(1) );
 
         let opath = "testData/output_index_test";
         mapper.write_index( opath.to_string() ).unwrap();
@@ -78,5 +109,6 @@ mod tests {
         assert_eq!(  mapper.names_store, mapper2.names_store );
 
     }
+
 
 }